Stick firestore export function in own src file
